# Remove the member variable of `jac_to_global` from base stepper