This also gives some boost in the event throughput

```
Main

CUDA unsync propagation/8   34369115636 ns   34355170928 ns            1 TracksPropagated=1.86289/s
CUDA unsync propagation/16     3289075 ns      3288918 ns          200 TracksPropagated=77.8371k/s
CUDA unsync propagation/32     4034031 ns      4033801 ns          173 TracksPropagated=253.855k/s
CUDA unsync propagation/64     5552317 ns      5552233 ns          124 TracksPropagated=737.721k/s
CUDA unsync propagation/128   12789134 ns     12789108 ns           54 TracksPropagated=1.28109M/s
CUDA unsync propagation/256   45228177 ns     45226861 ns           16 TracksPropagated=1.44905M/s
CUDA sync propagation/8        2092650 ns      2092419 ns          331 TracksPropagated=30.5866k/s
CUDA sync propagation/16       3308820 ns      3308826 ns          214 TracksPropagated=77.3688k/s
CUDA sync propagation/32       4082296 ns      4081955 ns          171 TracksPropagated=250.86k/s
CUDA sync propagation/64       5606346 ns      5606231 ns          123 TracksPropagated=730.616k/s
CUDA sync propagation/128     12890556 ns     12890330 ns           54 TracksPropagated=1.27103M/s
CUDA sync propagation/256     45122773 ns     45121536 ns           16 TracksPropagated=1.45243M/s

PR

CUDA unsync propagation/8   34369514205 ns   34366671308 ns            1 TracksPropagated=1.86227/s
CUDA unsync propagation/16     3150501 ns      3150404 ns          183 TracksPropagated=81.2594k/s
CUDA unsync propagation/32     3881827 ns      3881778 ns          180 TracksPropagated=263.797k/s
CUDA unsync propagation/64     4972935 ns      4972827 ns          139 TracksPropagated=823.676k/s
CUDA unsync propagation/128   11461089 ns     11460440 ns           60 TracksPropagated=1.42961M/s
CUDA unsync propagation/256   40707142 ns     40704544 ns           17 TracksPropagated=1.61004M/s
CUDA sync propagation/8        2017469 ns      2017485 ns          344 TracksPropagated=31.7227k/s
CUDA sync propagation/16       3198663 ns      3198392 ns          221 TracksPropagated=80.0402k/s
CUDA sync propagation/32       3917954 ns      3917836 ns          178 TracksPropagated=261.369k/s
CUDA sync propagation/64       4996937 ns      4996657 ns          138 TracksPropagated=819.748k/s
CUDA sync propagation/128     11514485 ns     11514296 ns           60 TracksPropagated=1.42293M/s
CUDA sync propagation/256     40804114 ns     40800854 ns           17 TracksPropagated=1.60624M/s
```